+++ src/lib.rs
@@ -2,6 +2,7 @@ mod processor;
 
 use aws_lambda_runtime_proxy::{LambdaRuntimeApiClient, MockLambdaRuntimeApiServer};
 use chrono::Utc;
+use http::{HeaderMap, HeaderValue};
 use tokio::{
   io::{stdin, AsyncBufReadExt, AsyncRead, BufReader},
   sync::{mpsc, oneshot},
@@ -112,31 +113,46 @@ impl<P> LogProxy<P> {
   where
     P: Processor,
   {
-    let checker_tx = self
+    let tx = self
       .processor
       .map(|p| spawn_reader(stdin(), p, self.buffer_size));
 
     MockLambdaRuntimeApiServer::bind(self.port)
       .await
       .serve(move |req| {
-        let checker_tx = checker_tx.clone();
+        let tx = tx.clone();
         async move {
-          if req.uri().path() == "/2018-06-01/runtime/invocation/next" {
-            // in lambda, send `invocation/next` will freeze current execution environment,
-            // unprocessed logs might be lost,
-            // so before proceeding, wait for the processors to finish processing the logs
+          match tx {
+            // if no processor, just forward the request
+            None => LambdaRuntimeApiClient::forward(req).await,
+            Some((checker_tx, next_tx)) => {
+              let is_invocation_next = req.uri().path() == "/2018-06-01/runtime/invocation/next";
 
-            // send checkers to reader threads
-            if let Some(checker_tx) = checker_tx {
-              let (ack_tx, ack_rx) = oneshot::channel();
-              checker_tx.send(ack_tx).await.unwrap();
-              // wait for the all checkers to finish
-              ack_rx.await.unwrap();
+              if is_invocation_next {
+                // in lambda, send `invocation/next` will freeze current execution environment,
+                // unprocessed logs might be lost,
+                // so before proceeding, wait for the processors to finish processing the logs
+
+                // send checkers to reader threads
+                let (ack_tx, ack_rx) = oneshot::channel();
+                checker_tx.send(ack_tx).await.unwrap();
+                // wait for the checker to finish
+                ack_rx.await.unwrap();
+              }
+
+              // forward the request to the real lambda runtime API, consume the request
+              let res = LambdaRuntimeApiClient::forward(req).await.unwrap();
+
+              if is_invocation_next {
+                let (ack_tx, ack_rx) = oneshot::channel();
+                next_tx.send((ack_tx, res.headers().clone())).await.unwrap();
+                // wait for the checker to finish
+                ack_rx.await.unwrap();
+              }
+
+              Ok(res)
             }
           }
-
-          // forward the request to the real lambda runtime API, consume the request
-          LambdaRuntimeApiClient::forward(req).await
         }
       })
       .await
@@ -147,11 +163,15 @@ fn spawn_reader<F: AsyncRead + Send + 'static, P: Processor + 'static>(
   file: F,
   mut processor: P,
   buffer_size: usize,
-) -> mpsc::Sender<oneshot::Sender<()>>
+) -> (
+  mpsc::Sender<oneshot::Sender<()>>,
+  mpsc::Sender<(oneshot::Sender<()>, HeaderMap<HeaderValue>)>,
+)
 where
   BufReader<F>: Unpin,
 {
   let (checker_tx, mut checker_rx) = mpsc::channel::<oneshot::Sender<()>>(1);
+  let (next_tx, mut next_rx) = mpsc::channel::<(oneshot::Sender<()>, HeaderMap<HeaderValue>)>(1);
   let (buffer_tx, mut buffer_rx) = mpsc::channel(buffer_size);
 
   // the reader thread, read from the file then push into the buffer
@@ -187,6 +207,11 @@ where
           // if we are writing to stdout, it is already line-buffered and will flush by line.
           // if we are writing to telemetry log fd, timestamp is appended so we don't need to flush it immediately.
         }
+        next = next_rx.recv() => {
+          let (ack, headers) = next.unwrap();
+          processor.next(headers).await;
+          ack.send(()).unwrap();
+        }
         // the server thread requests to check if the processor has finished processing the logs.
         checker = checker_rx.recv() => {
           // since we are using `biased` select, we don't need to check if there is a message in the buffer,
@@ -205,7 +230,7 @@ where
     }
   });
 
-  checker_tx
+  (checker_tx, next_tx)
 }
 
 #[cfg(test)]